+++ lib/dome.rb
@@ -1,9 +1,12 @@
+# frozen_string_literal: true
+
 require 'trollop'
 require 'aws-sdk'
 require 'colorize'
 require 'fileutils'
 require 'yaml'
 require 'hiera'
+require 'aws_assume_role'
 
 require 'dome/settings'
 require 'dome/version'