+++ lib/dome/environment.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Dome
   class Environment
     attr_reader :environment, :account, :ecosystem, :settings
@@ -25,18 +27,40 @@ module Dome
 
     def unset_aws_keys
       puts 'Unsetting environment variables '\
-        "#{'AWS_ACCESS_KEY'.colorize(:green)} and #{'AWS_SECRET_KEY'.colorize(:green)}"
+        "#{'AWS_ACCESS_KEY'.colorize(:green)}, #{'AWS_SECRET_KEY'.colorize(:green)}, "\
+        "#{'AWS_ACCESS_KEY_ID'.colorize(:green)}, #{'AWS_SECRET_ACCESS_KEY'.colorize(:green)}"\
+        " and #{'AWS_SESSION_TOKEN'.colorize(:green)}."
       ENV['AWS_ACCESS_KEY'] = nil
       ENV['AWS_SECRET_KEY'] = nil
+      ENV['AWS_SECRET_ACCESS_KEY'] = nil
+      ENV['AWS_ACCESS_KEY_ID'] = nil
+      ENV['AWS_SESSION_TOKEN'] = nil
     end
 
-    def aws_credentials
-      puts "Setting environment variable #{'AWS_PROFILE'.colorize(:green)} to your "\
-        "'account' name: #{@account.colorize(:green)}"
-      ENV['AWS_PROFILE'] = @account
+    def export_aws_keys(assumed_role)
+      puts 'Exporting temporary credentials to environment variables '\
+      "#{'AWS_ACCESS_KEY_ID'.colorize(:green)}, #{'AWS_SECRET_ACCESS_KEY'.colorize(:green)}"\
+      " and #{'AWS_SESSION_TOKEN'.colorize(:green)}."
+      ENV['AWS_ACCESS_KEY_ID'] = assumed_role.credentials.access_key_id
+      ENV['AWS_SECRET_ACCESS_KEY'] = assumed_role.credentials.secret_access_key
+      ENV['AWS_SESSION_TOKEN'] = assumed_role.credentials.session_token
+
       puts "Setting environment variable #{'AWS_DEFAULT_REGION'.colorize(:green)} "\
         "to #{'eu-west-1'.colorize(:green)}"
       ENV['AWS_DEFAULT_REGION'] = 'eu-west-1' # should we let people override this? doubtful
+    end
+
+    def aws_credentials
+      puts "Attempting to assume the role defined by your profile for 'account' name: "\
+        "#{@account.colorize(:green)}."
+      role_opts = { profile: account, role_session_name: account, use_mfa: true }
+      begin
+        assumed_role = AwsAssumeRole::DefaultProvider.new(role_opts).resolve
+      rescue StandardError => e
+        raise "ERROR: Unable to assume role: #{e}".colorize(:red)
+      end
+
+      export_aws_keys(assumed_role)
     end
 
     def valid_account?(account_name)
@@ -63,7 +87,6 @@ module Dome
 
     private
 
-    # rubocop:disable Metrics/AbcSize
     def generic_error_message
       puts "The 'account' and 'environment' variables are assigned based on your current directory.\n".colorize(:red)
       puts "The expected directory structure is 'terraform/<account>/<environment>'\n".colorize(:red)
@@ -75,10 +98,10 @@ module Dome
            " (where 'project' is defined using the 'project' key in your itv.yaml."
       puts "The accounts you have defined are: #{accounts}."
       puts '============================================================================='
-      puts 'To fix your issue, try the following:'
-      puts '1. Set your .aws/config to one of the valid accounts above.'
-      puts '2. Ensure you are running this from the correct directory.'
-      puts '3. Update your itv.yaml with the required environments or project.'
+      puts "To fix your issue, try the following:\n"\
+        "1. Set your .aws/config to one of the valid accounts above.\n"\
+        "2. Ensure you are running this from the correct directory.\n"\
+        '3. Update your itv.yaml with the required environments or project.'
     end
   end
 end