+++ lib/dome/terraform.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Dome
   class Terraform
     include Dome::Shell
@@ -36,6 +38,8 @@ module Dome
       @environment.aws_credentials
       puts '----------------------------------------------------------------'
     end
+    # rubocop:enable Metrics/MethodLength
+    # rubocop:enable Metrics/AbcSize
 
     def plan
       delete_terraform_directory